#6 Refactor: 컨트롤러, 서비스 구조 변경 - Service-1(Controller 계층에서는 여러 Service단을 주입받을 수 있다.), Controller-1(Service 계층에서는 다른 도메인의 Repository에 접근을 하지 않는다. ) 코드 컨벤션 적용 - Service-3(Service 계층에서 엔티티의 변경 권한을 확인한다.) 코드 컨벤션 적용

+++ src/main/java/com/drinkeg/drinkeg/service/wineClassBookMarkService/WineClassBookMarkServiceImpl.java
@@ -5,13 +5,11 @@ import com.drinkeg.drinkeg.converter.WineClassBookMarkConverter;
 import com.drinkeg.drinkeg.domain.Member;
 import com.drinkeg.drinkeg.domain.WineClass;
 import com.drinkeg.drinkeg.domain.WineClassBookMark;
-import com.drinkeg.drinkeg.dto.WineClassBookMarkDTO.request.WineClassBookMarkRequestDTO;
 import com.drinkeg.drinkeg.dto.WineClassBookMarkDTO.response.WineClassBookMarkResponseDTO;
 import com.drinkeg.drinkeg.exception.GeneralException;
-import com.drinkeg.drinkeg.repository.MemberRepository;
 import com.drinkeg.drinkeg.repository.WineClassBookMarkRepository;
-import com.drinkeg.drinkeg.repository.WineClassRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,18 +19,12 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class WineClassBookMarkServiceImpl implements WineClassBookMarkService{
     private final WineClassBookMarkRepository wineClassBookMarkRepository;
-    private final WineClassRepository wineClassRepository;
-    private final MemberRepository memberRepository;
 
     @Override
-    public WineClassBookMarkResponseDTO saveWineClassBookMark(Long userId, WineClassBookMarkRequestDTO wineClassBookMarkRequestDTO) {
-        if (wineClassBookMarkRepository.existsByMemberIdAndWineClassId(userId, wineClassBookMarkRequestDTO.getWineClassId()))
+    public WineClassBookMarkResponseDTO saveWineClassBookMark(Member member, WineClass wineClass) {
+        // 와인클래스 북마크 중복 확인
+        if (wineClassBookMarkRepository.existsByMemberAndWineClass(member, wineClass))
             throw new GeneralException(ErrorStatus.WINE_CLASS_BOOKMARK_DUPLICATED);
-
-        WineClass wineClass = wineClassRepository.findById(wineClassBookMarkRequestDTO.getWineClassId())
-                .orElseThrow(() -> new GeneralException(ErrorStatus.WINE_CLASS_NOT_FOUND));
-        Member member = memberRepository.findById(userId)
-                .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
 
         WineClassBookMark wineClassBookMark = WineClassBookMarkConverter.toWineClassBookMark(wineClass, member);
         wineClassBookMarkRepository.save(wineClassBookMark);
@@ -41,20 +33,21 @@ public class WineClassBookMarkServiceImpl implements WineClassBookMarkService{
     }
 
     @Override
-    public List<WineClassBookMarkResponseDTO> getWineClassBookMarksByUserId(Long userId) {
-        if (!memberRepository.existsById(userId))
-            throw new GeneralException(ErrorStatus.MEMBER_NOT_FOUND);
-
-        return wineClassBookMarkRepository.findAllByMemberId(userId)
+    public List<WineClassBookMarkResponseDTO> getWineClassBookMarksByUserId(Member member) {
+        return wineClassBookMarkRepository.findAllByMember(member)
                 .stream()
                 .map(WineClassBookMarkConverter::toWineClassBookMarkResponseDTO)
                 .collect(Collectors.toList());
     }
 
     @Override
-    public void deleteWineClassBookMarkById(Long wineClassBookMarkId) {
-        wineClassBookMarkRepository.findById(wineClassBookMarkId)
+    public void deleteWineClassBookMarkById(Member member, Long wineClassBookMarkId) {
+        WineClassBookMark wineClassBookMark = wineClassBookMarkRepository.findById(wineClassBookMarkId)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.WINE_CLASS_BOOKMARK_NOT_FOUND));
+        // 유저 권한 확인
+        if (!wineClassBookMark.getMember().equals(member))
+            throw new GeneralException(ErrorStatus.WINE_CLASS_BOOKMARK_UNAUTHORIZED);
+
         wineClassBookMarkRepository.deleteById(wineClassBookMarkId);
     }
 }
